alx_travel_app/requirement.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=3.15
drf-yasg>=1.21
django-environ>=0.10.0
mysqlclient>=2.1
celery>=5.3
